Crate nokhwa_bindings_linux

Structs

V4LCaptureDevice

The backend struct that interfaces with V4L2. To see what this does, please see CaptureBackendTrait.

Functions

id_to_known_camera_control

Attempts to convert a u32 V4L2 Control ID into a KnownCameraControl If the associated control is not found, this will return None (ColorEnable, Roll)

known_camera_control_to_id

Attempts to convert a KnownCameraControl into a V4L2 Control ID. If the associated control is not found, this will return None (ColorEnable, Roll)

query

query v4l2 cameras